Move vertices out of OpenGL namespace
