+++ osu.Framework/Graphics/OpenGL/Batches/QuadBatch.cs
@@ -5,7 +5,7 @@
 
 using System;
 using osu.Framework.Graphics.OpenGL.Buffers;
-using osu.Framework.Graphics.OpenGL.Vertices;
+using osu.Framework.Graphics.Rendering.Vertices;
 using osuTK.Graphics.ES30;
 
 namespace osu.Framework.Graphics.OpenGL.Batches
