+++ osu.Framework/Graphics/OpenGL/Batches/VertexBatch.cs
@@ -7,8 +7,8 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using osu.Framework.Graphics.OpenGL.Buffers;
-using osu.Framework.Graphics.OpenGL.Vertices;
 using osu.Framework.Graphics.Rendering;
+using osu.Framework.Graphics.Rendering.Vertices;
 using osu.Framework.Statistics;
 
 namespace osu.Framework.Graphics.OpenGL.Batches
