+++ osu.Framework/Graphics/OpenGL/Buffers/LinearVertexBuffer.cs
@@ -4,7 +4,7 @@
 #nullable disable
 
 using System;
-using osu.Framework.Graphics.OpenGL.Vertices;
+using osu.Framework.Graphics.Rendering.Vertices;
 using osuTK.Graphics.ES30;
 
 namespace osu.Framework.Graphics.OpenGL.Buffers
