+++ osu.Framework/Graphics/OpenGL/Buffers/QuadVertexBuffer.cs
@@ -5,8 +5,8 @@
 
 using System;
 using System.Diagnostics;
-using osu.Framework.Graphics.OpenGL.Vertices;
 using osu.Framework.Graphics.Rendering;
+using osu.Framework.Graphics.Rendering.Vertices;
 using osuTK.Graphics.ES30;
 
 namespace osu.Framework.Graphics.OpenGL.Buffers
