+++ osu.Framework/Graphics/OpenGL/Buffers/VertexBuffer.cs
@@ -5,10 +5,10 @@
 
 using System;
 using System.Buffers;
-using osu.Framework.Graphics.OpenGL.Vertices;
 using osuTK.Graphics.ES30;
 using osu.Framework.Statistics;
 using osu.Framework.Development;
+using osu.Framework.Graphics.Rendering.Vertices;
 using SixLabors.ImageSharp.Memory;
 
 namespace osu.Framework.Graphics.OpenGL.Buffers
