+++ osu.Framework/Graphics/OpenGL/Buffers/VertexUtils.cs
@@ -10,9 +10,10 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Reflection;
 using System.Runtime.InteropServices;
+using osu.Framework.Graphics.Rendering.Vertices;
 using osuTK.Graphics.ES30;
 
-namespace osu.Framework.Graphics.OpenGL.Vertices
+namespace osu.Framework.Graphics.OpenGL.Buffers
 {
     /// <summary>
     /// Helper method that provides functionality to enable and bind vertex attributes.
