+++ osu.Framework/Graphics/Rendering/Vertices/DepthWrappingVertex.cs
@@ -7,7 +7,7 @@ using System;
 using System.Runtime.InteropServices;
 using osuTK.Graphics.ES30;
 
-namespace osu.Framework.Graphics.OpenGL.Vertices
+namespace osu.Framework.Graphics.Rendering.Vertices
 {
     [StructLayout(LayoutKind.Sequential)]
     internal struct DepthWrappingVertex<TVertex> : IVertex, IEquatable<DepthWrappingVertex<TVertex>>
