+++ osu.Framework/Graphics/Rendering/Vertices/ParticleVertex2D.cs
@@ -9,7 +9,7 @@ using osuTK;
 using osuTK.Graphics;
 using osuTK.Graphics.ES30;
 
-namespace osu.Framework.Graphics.OpenGL.Vertices
+namespace osu.Framework.Graphics.Rendering.Vertices
 {
     [StructLayout(LayoutKind.Sequential)]
     public struct ParticleVertex2D : IEquatable<ParticleVertex2D>, IVertex
