+++ osu.Framework/Graphics/UserInterface/CircularProgressDrawNode.cs
@@ -9,8 +9,8 @@ using System;
 using osu.Framework.Graphics.Primitives;
 using osuTK.Graphics;
 using osu.Framework.Extensions.MatrixExtensions;
-using osu.Framework.Graphics.OpenGL.Vertices;
 using osu.Framework.Graphics.Rendering;
+using osu.Framework.Graphics.Rendering.Vertices;
 
 namespace osu.Framework.Graphics.UserInterface
 {
